feat : placeholder 및 에디터 문장별 padding 여유있게 추가
