+++ src/components/editor/utils/parseMarkdown.js
@@ -1,56 +1,12 @@
-import {DUMMY_CHARACTER, SPACEBAR_KEY_CODE, ENTER_KEY_CODE, BACKTICK_KEY_CODE} from "../constants/constants.js"
+import {DUMMY_CHARACTER, SPACEBAR_KEY, ENTER_KEY, BACKTICK_KEY} from "../../../constants/constants.js"
+import insertElementInEditor from "./insertElementInEditor.js";
 
 const selection = window.getSelection();
 let offset = null;
 
-export function insertElementInEditor($element){
-    const range = selection.getRangeAt(0);
-    range.selectNode(selection.anchorNode);
-    range.deleteContents();
-    range.insertNode($element);
-    range.setStartAfter($element);
-    range.setEndAfter($element);
-    selection.removeAllRanges();
-    selection.addRange(range);
-    range.detach();
-}
-
-export function onPressTab(){
-  const $editor = document.querySelector("[name=content]");
-  const $doc = $editor.ownerDocument.defaultView;
-  const docSelection = $doc.getSelection();
-  const range = docSelection.getRangeAt(0);
-  const tabNode = document.createTextNode("\u00a0\u00a0\u00a0\u00a0");
-  range.insertNode(tabNode);
-  range.setStartAfter(tabNode);
-  range.setEndAfter(tabNode);
-  docSelection.removeAllRanges();
-  docSelection.addRange(range);
-  return;
-}
-
-export function onPressInCodeBlock(e){
-  const $currentParent = selection.anchorNode.parentElement;
-  const $current = selection.anchorNode;
-  if($currentParent.classList.contains("code-block") && 
-  selection.anchorOffset === $current.length){
-    e.preventDefault();
-    const range = selection.getRangeAt(0);
-    const $dummy = document.createTextNode(DUMMY_CHARACTER);
-    range.setStartAfter($currentParent);
-    range.setEndAfter($currentParent);
-    range.insertNode($dummy);
-    range.setStart($dummy, 0);
-    range.setEnd($dummy, 1);
-    selection.removeAllRanges();
-    selection.addRange(range);
-    range.detach();
-  }
-}
-
-export function parseMarkdown(key){
+export default function parseMarkdown(key){
   const line = selection.anchorNode.textContent;
-  if(key === SPACEBAR_KEY_CODE){
+  if(key === SPACEBAR_KEY){
     if(line.match(/^\#/g)){
       const hashTagCount = (line.match(/#/g)).length;
       if(hashTagCount > 3) return;
@@ -60,7 +16,7 @@ export function parseMarkdown(key){
       insertElementInEditor($header);
       $header.textContent = line.substring(hashTagCount) + DUMMY_CHARACTER;
     }
-  }else if(key === ENTER_KEY_CODE){
+  }else if(key === ENTER_KEY){
     const range = selection.getRangeAt(0);
     const $br = document.createElement("br");
     if(range.commonAncestorContainer.nodeName === "SPAN"){
@@ -79,7 +35,7 @@ export function parseMarkdown(key){
       $newDiv.innerHTML = `<br>`;
       $parentDiv.replaceChild($newDiv, $deleteDiv);
     }
-  }else if(key === BACKTICK_KEY_CODE){
+  }else if(key === BACKTICK_KEY){
     const range = selection.getRangeAt(0);
     let $divParent = range.commonAncestorContainer.parentElement;
     while($divParent.nodeName !== "DIV"){
