+++ src/constants/constants.js
@@ -2,11 +2,11 @@ export const DUMMY_CHARACTER = "\u00a0";
 
 export const API_WAIT_TIME = 1500;
 
-export const SPACEBAR_KEY_CODE = 32;
-export const RIGHT_ARROW_KEY_CODE = 39;
-export const BACKTICK_KEY_CODE = 192;
-export const ENTER_KEY_CODE = 13;
-export const TAB_KEY_CODE = 9;
+export const SPACEBAR_KEY = " ";
+export const RIGHT_ARROW_KEY = "ArrowRight";
+export const BACKTICK_KEY = "`";
+export const ENTER_KEY = "Enter";
+export const TAB_KEY = "Tab";
 
 export const ROUTE_EVENT_NAME  = "route-change";
 export const ROUTE_REPLACE_EVENT_NAME = "route-replace";
